+++ src/dftbp/dftbplus/main.F90
@@ -25,7 +25,6 @@ module dftbp_dftbplus_main
   use dftbp_dftb_determinants, only : TDftbDeterminants, TDftbDeterminants_init, determinants
   use dftbp_dftb_dftbplusu, only : TDftbU
   use dftbp_dftb_dispersions, only : TDispersionIface
-  use dftbp_dftb_elstatpot, only : TElStatPotentials
   use dftbp_dftb_energytypes, only : TEnergies
   use dftbp_dftb_etemp, only : electronFill, Efilling
   use dftbp_dftb_extfields, only : addUpExternalField
@@ -48,7 +47,7 @@ module dftbp_dftbplus_main
   use dftbp_dftb_rangeseparated, only : TRangeSepFunc
   use dftbp_dftb_repulsive_repulsive, only : TRepulsive
   use dftbp_dftb_scc, only : TScc
-  use dftbp_dftb_shift, only : addShift
+  use dftbp_dftb_shift, only : addShift, addAtomicMultipoleShift
   use dftbp_dftb_slakocont, only : TSlakoCont
   use dftbp_dftb_sparse2dense, only : unpackHPauli, unpackHS, blockSymmetrizeHS, packHS,&
       & blockSymmetrizeHS, packHS, SymmetrizeHS, unpackHelicalHS, packerho, blockHermitianHS,&
@@ -1279,7 +1278,7 @@ contains
             & this%filling, this%rhoPrim, this%xi, this%orbitalL, this%HSqrReal,&
             & this%SSqrReal, this%eigvecsReal, this%iRhoPrim, this%HSqrCplx, this%SSqrCplx,&
             & this%eigvecsCplx, this%rhoSqrReal, this%deltaRhoInSqr, this%deltaRhoOutSqr,&
-            & this%nNeighbourLC, this%deltaDftb, this%apiCallBack, errStatus)
+            & this%nNeighbourLC, this%deltaDftb, errStatus, this%apiCallBack)
         if (errStatus%hasError()) then
           call error(errStatus%message)
         end if
@@ -1452,11 +1451,9 @@ contains
       call getDipoleMoment(this%qOutput, this%q0, this%multipoleOut%dipoleAtom, this%coord0,&
           & this%dipoleMoment(:,this%deltaDftb%iDeterminant), this%iAtInCentralRegion)
     #:block DEBUG_CODE
-      if (this%hamiltonianType == hamiltonianTypes%dftb) then
-        call checkDipoleViaHellmannFeynman(env, this%rhoPrim, this%q0, this%coord0, this%ints,&
-            & this%orb, this%neighbourList, this%nNeighbourSk, this%species, this%iSparseStart,&
-            & this%img2CentCell, this%eFieldScaling)
-      end if
+      call checkDipoleViaHellmannFeynman(env, this%rhoPrim, this%q0, this%coord0, this%ints,&
+          & this%orb, this%neighbourList, this%nNeighbourSk, this%species, this%iSparseStart,&
+          & this%img2CentCell, this%eFieldScaling, this%hamiltonianType, this%nDipole)
     #:endblock DEBUG_CODE
     end if
 
@@ -2426,7 +2423,7 @@ contains
       & tFixEf, tMulliken, iDistribFn, tempElec, nEl, parallelKS, Ef, mu, energy, rangeSep, eigen,&
       & filling, rhoPrim, xi, orbitalL, HSqrReal, SSqrReal, eigvecsReal, iRhoPrim, HSqrCplx,&
       & SSqrCplx, eigvecsCplx, rhoSqrReal, deltaRhoInSqr, deltaRhoOutSqr, nNeighbourLC, deltaDftb,&
-      & apiCallBack, errStatus)
+      & errStatus, apiCallBack)
 
     !> Environment settings
     type(TEnvironment), intent(inout) :: env
@@ -2585,6 +2582,9 @@ contains
     !> Status of operation
     type(TStatus), intent(out) :: errStatus
 
+    !> Object for invocation of the density, overlap, and hamiltonian matrices exported by callbacks
+    type(TAPICallback), intent(inout), optional :: apiCallBack
+
     !! Number of spin channels
     integer :: nSpin
 
@@ -2625,8 +2625,8 @@ contains
           & tSpinSharedEf, tSpinOrbit, tDualSpinOrbit, tFillKSep, tFixEf, tMulliken, iDistribFn,&
           & tempElec, nEl, parallelKS, Ef, energy, rangeSep, eigen, filling, rhoPrim, xi,&
           & orbitalL, HSqrReal, SSqrReal, eigvecsReal, iRhoPrim, HSqrCplx, SSqrCplx, eigvecsCplx,&
-          & rhoSqrReal, deltaRhoInSqr, deltaRhoOutSqr, nNeighbourLC, deltaDftb, apiCallBack,&
-          & iSCC, errStatus)
+          & rhoSqrReal, deltaRhoInSqr, deltaRhoOutSqr, nNeighbourLC, deltaDftb, iSCC, errStatus,&
+          & apiCallBack)
       @:PROPAGATE_ERROR(errStatus)
 
     case(electronicSolverTypes%omm, electronicSolverTypes%pexsi, electronicSolverTypes%ntpoly,&
@@ -2653,7 +2653,7 @@ contains
       & tFillKSep, tFixEf, tMulliken, iDistribFn, tempElec, nEl, parallelKS, Ef, energy, rangeSep,&
       & eigen, filling, rhoPrim, xi, orbitalL, HSqrReal, SSqrReal, eigvecsReal, iRhoPrim,&
       & HSqrCplx, SSqrCplx, eigvecsCplx, rhoSqrReal, deltaRhoInSqr, deltaRhoOutSqr, nNeighbourLC,&
-      & deltaDftb, apiCallBack, iSCC, errStatus)
+      & deltaDftb, iSCC, errStatus, apiCallBack)
 
     !> Environment settings
     type(TEnvironment), intent(inout) :: env
@@ -2800,15 +2800,15 @@ contains
     !> Determinant derived type
     type(TDftbDeterminants), intent(inout) :: deltaDftb
 
-    !> Object for invocation of the density, overlap, and hamiltonian matrices exported by callbacks
-    type(Tapicallback), intent(inout) :: apiCallBack
-
     !> SCC iteration index. Used here to determine if we need invoke overlap and hamiltonian
     !> matrices exporting callbacks now.
     integer, intent(in) :: iSCC
 
     !> Status of operation
     type(TStatus), intent(out) :: errStatus
+
+    !> Object for invocation of the density, overlap, and hamiltonian matrices exported by callbacks
+    type(Tapicallback), intent(inout), optional :: apiCallBack
 
     integer :: nSpin
 
@@ -2819,13 +2819,13 @@ contains
         call buildAndDiagDenseRealHam(env, denseDesc, ints, species, neighbourList, nNeighbourSK,&
             & iSparseStart, img2CentCell, orb, tHelical, coord, electronicSolver, parallelKS,&
             & rangeSep, deltaRhoInSqr, nNeighbourLC, HSqrReal, SSqrReal, eigVecsReal, eigen(:,1,:),&
-            & apiCallBack, iSCC, errStatus)
+            & iSCC, errStatus, apiCallBack)
         @:PROPAGATE_ERROR(errStatus)
       else
         call buildAndDiagDenseCplxHam(env, denseDesc, ints, kPoint, neighbourList,&
             & nNeighbourSK, iSparseStart, img2CentCell, iCellVec, cellVec, electronicSolver,&
             & parallelKS, tHelical, orb, species, coord, HSqrCplx, SSqrCplx, eigVecsCplx, eigen,&
-            & apiCallBack, iSCC, errStatus)
+            & iSCC, errStatus, apiCallBack)
         @:PROPAGATE_ERROR(errStatus)
       end if
     else
@@ -2868,8 +2868,8 @@ contains
   !> Builds and diagonalises dense Hamiltonians.
   subroutine buildAndDiagDenseRealHam(env, denseDesc, ints, species, neighbourList, nNeighbourSK,&
       & iSparseStart, img2CentCell, orb, tHelical, coord, electronicSolver, parallelKS, rangeSep,&
-      & deltaRhoInSqr, nNeighbourLC, HSqrReal, SSqrReal, eigvecsReal, eigen, apiCallBack, iSCC,&
-      & errStatus)
+      & deltaRhoInSqr, nNeighbourLC, HSqrReal, SSqrReal, eigvecsReal, eigen, iSCC, errStatus,&
+      & apiCallBack)
 
     !> Environment settings
     type(TEnvironment), intent(inout) :: env
@@ -2932,18 +2932,17 @@ contains
     !> eigenvalues
     real(dp), intent(out) :: eigen(:,:)
 
-    !> Object for invocation of the density, overlap, and hamiltonian matrices exported by callbacks
-    type(Tapicallback), intent(in), optional :: apiCallBack
-
     !> SCC iteration index. Used here to determine if we need invoke overlap and hamiltonian
     !> matrices exporting callbacks now.
     integer, intent(in) :: iSCC
 
-    integer :: iK, iKS, iSpin
-
     !> Status of operation
     type(TStatus), intent(out) :: errStatus
 
+    !> Object for invocation of the density, overlap, and hamiltonian matrices exported by callbacks
+    type(Tapicallback), intent(in), optional :: apiCallBack
+
+    integer :: iK, iKS, iSpin
 
     eigen(:,:) = 0.0_dp
     do iKS = 1, parallelKS%nLocalKS
@@ -3028,8 +3027,8 @@ contains
   !> Builds and diagonalises dense k-point dependent Hamiltonians.
   subroutine buildAndDiagDenseCplxHam(env, denseDesc, ints, kPoint, neighbourList,&
       & nNeighbourSK, iSparseStart, img2CentCell, iCellVec, cellVec, electronicSolver, parallelKS,&
-      & tHelical, orb, species, coord, HSqrCplx, SSqrCplx, eigvecsCplx, eigen, apiCallBack,&
-      & iSCC, errStatus)
+      & tHelical, orb, species, coord, HSqrCplx, SSqrCplx, eigvecsCplx, eigen, iSCC, errStatus,&
+      & apiCallBack)
 
     !> Environment settings
     type(TEnvironment), intent(inout) :: env
@@ -3091,15 +3090,15 @@ contains
     !> eigenvalues
     real(dp), intent(out) :: eigen(:,:,:)
 
-    !> Object for invocation of the density, overlap, and hamiltonian matrices exporting callbacks
-    type(Tapicallback), intent(in), optional :: apiCallBack
-
     !> SCC iteration index. Used here to determine if we need invoke overlap and hamiltonian
     !> matrices exporting callbacks now.
     integer, intent(in) :: iSCC
 
     !> Status of operation
     type(TStatus), intent(out) :: errStatus
+
+    !> Object for invocation of the density, overlap, and hamiltonian matrices exporting callbacks
+    type(Tapicallback), intent(in), optional :: apiCallBack
 
     integer :: iKS, iK, iSpin
 
@@ -3485,7 +3484,6 @@ contains
 
     !> Object for invocation of the density, overlap, and hamiltonian matrices exporting callbacks
     type(Tapicallback), intent(in), optional :: apiCallBack
-
 
     integer :: iKS, iK, iSpin
 
@@ -4773,12 +4771,13 @@ contains
       end do
     end if
 
+
   end subroutine getDipoleMoment
 
 
   !> Prints dipole moment calculated by the derivative of H with respect to the external field.
   subroutine checkDipoleViaHellmannFeynman(env, rhoPrim, q0, coord0, ints, orb, neighbourList,&
-      & nNeighbourSK, species, iSparseStart, img2CentCell, eFieldScaling)
+      & nNeighbourSK, species, iSparseStart, img2CentCell, eFieldScaling, iHamiltonianType, nDipole)
 
     !> Environment settings
     type(TEnvironment), intent(in) :: env
@@ -4816,7 +4815,14 @@ contains
     !> Instance of electric/dipole scaling due to any dielectric media effects
     class(TScaleExtEField), intent(in) :: eFieldScaling
 
+    !> Hamiltonian type
+    integer, intent(in) :: iHamiltonianType
+
+    !> Number of atomic dipole moment components
+    integer, intent(in) :: nDipole
+
     real(dp), allocatable :: hprime(:,:), dipole(:,:), potentialDerivative(:,:)
+    real(dp), allocatable :: potentialGradDeriv(:,:)
     integer :: nAtom, sparseSize, iAt, iCart
 
     sparseSize = size(ints%overlap)
@@ -4827,19 +4833,36 @@ contains
     write(stdOut,*)
     write(stdOut, "(A)", advance='no') 'Hellmann Feynman dipole:'
 
+  #:block DEBUG_CODE
+    if (nDipole > 0) then
+      @:ASSERT(iHamiltonianType == hamiltonianTypes%xtb)
+      allocate(potentialGradDeriv(nDipole, nAtom))
+    end if
+  #:endblock DEBUG_CODE
+
     ! loop over directions
     do iCart = 1, 3
       potentialDerivative(:,:) = 0.0_dp
       ! Potential from dH/dE
       potentialDerivative(:,1) = -eFieldScaling%scaledExtEField(coord0(iCart,:))
       hprime(:,:) = 0.0_dp
-      dipole(:,:) = 0.0_dp
-      call addShift(env, hprime, ints%overlap, nNeighbourSK, neighbourList%iNeighbour, species, orb,&
-          & iSparseStart, nAtom, img2CentCell, potentialDerivative, .true.)
 
+      call addShift(env, hprime, ints%overlap, nNeighbourSK, neighbourList%iNeighbour, species,&
+          & orb, iSparseStart, nAtom, img2CentCell, potentialDerivative, .true.)
+
+      if (nDipole > 0) then
+        potentialGradDeriv(:,:) = 0.0_dp
+        potentialGradDeriv(iCart,:) = -eFieldScaling%scaledExtEField(1.0_dp)
+
+        call addAtomicMultipoleShift(hPrime, ints%dipoleBra, ints%dipoleKet, nNeighbourSK, &
+            & neighbourList%iNeighbour, species, orb, iSparseStart, nAtom, img2CentCell, &
+            & potentialGradDeriv)
+      end if
+
+      dipole(:,:) = 0.0_dp
       ! evaluate <psi| dH/dE | psi> = Tr_part rho dH/dE
-      call mulliken(env, dipole, hprime(:,1), rhoPrim(:,1), orb, neighbourList%iNeighbour, nNeighbourSK,&
-          & img2CentCell, iSparseStart)
+      call mulliken(env, dipole, hprime(:,1), rhoPrim(:,1), orb, neighbourList%iNeighbour,&
+          & nNeighbourSK, img2CentCell, iSparseStart)
 
       ! add nuclei term for derivative wrt E
       do iAt = 1, nAtom