+++ src/dftbp/type/oldskdata.F90
@@ -13,14 +13,15 @@ module dftbp_type_oldskdata
   use dftbp_common_accuracy, only : dp, lc
   use dftbp_common_constants, only : amu__au
   use dftbp_common_file, only : TFileDescr, openFile, closeFile
-  use dftbp_dftb_rangeseparated, only : TRangeSepSKTag
+  use dftbp_dftb_rangeseparated, only : TRangeSepSKTag, rangeSepFunc
   use dftbp_dftb_repulsive_polyrep, only : TPolyRepInp
   use dftbp_dftb_repulsive_splinerep, only : TSplineRepInp
   use dftbp_io_message, only : error
+  use dftbp_io_charmanip, only : tolower
   implicit none
 
   private
-  public :: TOldSKData, readFromFile, readSplineRep
+  public :: TOldSKData, readFromFile, readSplineRep, inquireRangeSepTag
 
 
   !> Represents the Slater-Koster data in an SK file.
@@ -103,7 +104,7 @@ contains
     !> Repulsive polynomial part of the SK-file.
     type(TPolyRepInp), intent(out), optional :: polyRepIn
 
-    !> Reads rangeseparation parameter from SK file
+    !> Reads range-separated parameter(s) from SK file
     type(TRangeSepSKTag), intent(inout), optional :: rangeSepSK
 
     type(TFileDescr) :: file
@@ -151,7 +152,7 @@ contains
       read (file%unit,*, iostat=iostat) skData%mass, (coeffs(ii), ii = 2, 9), &
           &polyCutoff, rDummy, (rDummy, ii = 12, 20)
       call checkIoError(iostat, fileName, "Unable to read 3rd data line")
-      skData%mass = skData%mass * amu__au  !convert to atomic units
+      skData%mass = skData%mass * amu__au  ! convert to atomic units
     else
       read (file%unit,*, iostat=iostat) rDummy, (coeffs(ii), ii = 2, 9),&
           & polyCutoff, (rDummy, ii = 11, 20)
@@ -169,14 +170,14 @@ contains
     skData%skOver(:,:) = 0.0_dp
     do iGrid = 1, skData%nGrid
       if (tExtended) then
-        read (file%unit, *, iostat=iostat) &
-            &(skData%skHam(iGrid, ii), ii = 1, nSKInter), &
-            &(skData%skOver(iGrid, ii), ii = 1, nSKInter)
+        read (file%unit, *, iostat=iostat)&
+            & (skData%skHam(iGrid, ii), ii = 1, nSKInter),&
+            & (skData%skOver(iGrid, ii), ii = 1, nSKInter)
         call checkIoError(iostat, fileName, "Reading error for integrals")
       else
-        read(file%unit,*, iostat=iostat) &
-            &(skData%skHam(iGrid, iSKInterOld(ii)), ii = 1, nSKInterOld), &
-            &(skData%skOver(iGrid, iSKInterOld(ii)), ii = 1, nSKInterOld)
+        read(file%unit,*, iostat=iostat)&
+            & (skData%skHam(iGrid, iSKInterOld(ii)), ii = 1, nSKInterOld),&
+            & (skData%skOver(iGrid, iSKInterOld(ii)), ii = 1, nSKInterOld)
         call checkIoError(iostat, fileName, "Reading error for integrals")
       end if
     end do
@@ -188,14 +189,72 @@ contains
 
     call readSplineRep(file%unit, fileName, splineRepIn, iSp1, iSp2)
 
-    !> Read range separation parameter
+    ! Read range-separated parameter(s)
     if (present(rangeSepSK)) then
-       call readRangeSep(file%unit, fileName, rangeSepSK)
+      call readRangeSepParams(file%unit, fileName, rangeSepSK)
     end if
 
     call closeFile(file)
 
   end subroutine OldSKData_readFromFile
+
+
+  !> Reads range-separated parameter(s) from an open file.
+  subroutine readRangeSepParams(fp, fname, rangeSepSK)
+
+    !> File identifier
+    integer, intent(in) :: fp
+
+    !> File name
+    character(len=*), intent(in) :: fname
+
+    !> range-separated parameter(s)
+    type(TRangeSepSKTag), intent(inout) :: rangeSepSK
+
+    !! Error status
+    integer :: iErr
+
+    !! Temporary character storage
+    character(lc) :: strDummy
+
+    !! True, if range-separated extra tag was found in SK-file
+    logical :: isRangeSepTag
+
+    !! Extra tag in SK-files, defining the type of range-separated
+    integer :: rangeSepTag
+
+    call inquireRangeSepTag(fname, rangeSepTag, fp=fp)
+    rewind(fp)
+
+    ! Seek range-separated section in SK file
+    do
+      read(fp, '(A)', iostat=iErr) strDummy
+      if (iErr /= 0) then
+        isRangeSepTag = .false.
+        exit
+      elseif (strDummy == "RangeSep") then
+        isRangeSepTag = .true.
+        exit
+      end if
+    end do
+
+    if (.not. isRangeSepTag) then
+      write(strDummy, "(A,A,A)") "Range-separated calculation requested, but SK-file '",&
+          & trim(fname), "' is not a suitable parametrization."
+      call error(strDummy)
+    end if
+
+    if (rangeSepTag == rangeSepFunc%lc) then
+      read(fp, *, iostat=iErr) strDummy, rangeSepSK%omega
+      call checkioerror(iErr, fname, "Error in reading range-separated parameter(s)")
+    end if
+
+    if (rangeSepSK%omega < 0.0_dp) then
+      write(strDummy, "(A)") "Range-separation parameter is negative."
+      call error(strDummy)
+    end if
+
+  end subroutine readRangeSepParams
 
 
   !> Reads the repulsive from an open file.
@@ -268,64 +327,79 @@ contains
   end subroutine OldSKData_readsplinerep
 
 
-  !> Reads the RangeSep data from an open file.
-  subroutine readRangeSep(fp, fname, rangeSepSK)
-
-    !> File identifier
-    integer, intent(in) :: fp
+  !> Inquires range-separated extra tag of SK-file.
+  subroutine inquireRangeSepTag(fname, rangeSepTag, fp)
 
     !> File name
-    character(*), intent(in) :: fname
+    character(len=*), intent(in) :: fname
 
-    !> Rangesep data
-    type(TRangeSepSKTag), intent(inout) :: rangeSepSK
+    !> Range-separated extra tag, if allocated
+    integer, intent(out) :: rangeSepTag
 
-    integer :: iostat
-    character(lc) :: chdummy
-    real(dp) :: omega
-    logical :: hasRangeSep
+    !> File identifier, if file is already open
+    integer, intent(in), optional :: fp
 
-    !> Seek rangesep part in SK file
+    !! File descriptor
+    type(TFileDescr) :: file
+
+    !! Error status
+    integer :: iErr
+
+    !! Temporary character storage
+    character(lc) :: strDummy
+
+    !! True, if range-separated extra tag was found in SK-file
+    logical :: isRangeSepTag
+
+    !> File identifier
+    integer :: fd
+
+    if (present(fp)) then
+      fd = fp
+    else
+      call openFile(file, fname, mode="r", ioStat=iErr)
+      fd = file%unit
+      call checkIoError(iErr, fname, "Unable to open file")
+    end if
+
+    ! Seek range-separated extra tag in SK-file
     do
-      read(fp, '(A)', iostat=iostat) chdummy
-      if (iostat /= 0) then
-        hasRangeSep = .false.
+      read(fd, '(A)', iostat=iErr) strDummy
+      if (iErr /= 0) then
+        isRangeSepTag = .false.
         exit
-      elseif (chdummy == "RangeSep") then
-        hasRangeSep = .true.
+      elseif (strDummy == "RangeSep") then
+        isRangeSepTag = .true.
         exit
       end if
     end do
 
-    if ( .not. hasRangeSep) then
-      write(chdummy, "(A,A,A)") "RangeSep extension tag not found in file '",&
-          & trim(fname), "'"
-      call error(chdummy)
+    if (isRangeSepTag) then
+      read(fd, *, iostat=iErr) strDummy
+      call checkIoError(iErr, fname, "Error in reading range-separated extra tag and method")
+
+      select case(tolower(trim(strDummy)))
+      case ("lc")
+        rangeSepTag = rangeSepFunc%lc
+      case default
+        write(strDummy, "(A,A,A)") "Unknown range-separated method in SK-file '",&
+            & trim(fname), "'"
+        call error(strDummy)
+      end select
+    else
+      rangeSepTag = rangeSepFunc%none
     end if
 
-    read(fp, *, iostat=iostat) chdummy, omega
-    call checkioerror(iostat, fname, "Error in reading range-sep method and range-sep parameter")
+    if (.not. present(fp)) call closeFile(file)
 
-    if (chdummy /= "LC") then
-      write(chdummy, "(A,A,A)") "Unknown range-separation method in SK file '", trim(fname), "'"
-      call error(chdummy)
-    end if
-
-    if (omega < 0.0_dp) then
-      write(chdummy, "(A)") "Range-separation parameter is negative"
-      call error(chdummy)
-   end if
-
-   rangeSepSK%omega = omega
-
-  end subroutine ReadRangeSep
+  end subroutine inquireRangeSepTag
 
 
   !> Checks for IO errors and prints message.
-  subroutine checkIOError(iostat, fname, msg)
+  subroutine checkIOError(iErr, fname, msg)
 
     !> Flag of the IO operation.
-    integer, intent(in) :: iostat
+    integer, intent(in) :: iErr
 
     !> Name of the file.
     character(*), intent(in) :: fname
@@ -333,7 +407,7 @@ contains
     !> Message to print if IO operation flag is non-zero.
     character(*), intent(in) :: msg
 
-    if (iostat /= 0) then
+    if (iErr /= 0) then
       call error("IO error in file '" // trim(fname) // "': " // trim(msg))
     end if